+++ ydb/core/kqp/provider/yql_kikimr_gateway.h
@@ -511,6 +511,7 @@ struct TTtlSettings {
     struct TTier {
         TDuration ApplyAfter;
         std::optional<TString> StorageName;
+        std::optional<TString> ObjectKeyPrefix;
     };
 
     TString ColumnName;