+++ ydb/core/tx/columnshard/blobs_action/tier/write.cpp
@@ -1,4 +1,5 @@
 #include "write.h"
+#include "object_key.h"
 
 #include <ydb/core/tx/columnshard/columnshard_impl.h>
 #include <ydb/core/wrappers/events/common.h>
@@ -8,7 +9,7 @@
 namespace NKikimr::NOlap::NBlobOperations::NTier {
 
 void TWriteAction::DoSendWriteBlobRequest(const TString& data, const TUnifiedBlobId& blobId) {
-    auto awsRequest = Aws::S3::Model::PutObjectRequest().WithKey(blobId.GetLogoBlobId().ToString());
+    auto awsRequest = Aws::S3::Model::PutObjectRequest().WithKey(TObjectKey(GetStorageId()).Make(blobId.GetLogoBlobId()));
 
     TString moveData = data;
     auto request = std::make_unique<NWrappers::NExternalStorage::TEvPutObjectRequest>(awsRequest, std::move(moveData));
@@ -33,7 +34,7 @@ void TWriteAction::DoOnExecuteTxBeforeWrite(NColumnShard::TColumnShard& /*self*/
 }
 
 NKikimr::NOlap::TUnifiedBlobId TWriteAction::AllocateNextBlobId(const TString& data) {
-    return TUnifiedBlobId(Max<ui32>(), TLogoBlobID(TabletId, Generation, Step, TLogoBlobID::MaxChannel, data.size(), ++BlobIdsCounter));
+    return TUnifiedBlobId(Max<ui32>(), TLogoBlobID(TabletId, Generation, Step, TObjectKey(GetStorageId()).GetChannelForWriting(), data.size(), ++BlobIdsCounter));
 }
 
 void TWriteAction::DoOnCompleteTxAfterWrite(NColumnShard::TColumnShard& /*self*/, const bool blobsWroteSuccessfully) {